django/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
dj-database-url==2.1.0
Django==4.2.5
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
django-recaptcha3==0.4.0
djangorestframework==3.14.0
environs==9.5.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
packaging==23.1
phonenumbers==8.13.21
Pillow==10.0.1
pip-review==1.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
